removed unecessary navigation files, fixed created by field when making study group, improved follow/unfollow alert text
