+++ StudyMeets/App.js
@@ -1,26 +1,52 @@
 import React from 'react';
-import { TouchableOpacity, Text } from 'react-native';
+import { View, Text, StyleSheet } from 'react-native';
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import Login from './app/screens/Login';
 import Explore from './app/screens/Explore';
-import SettingsNavigation from './app/screens/SettingsNavigation';
+import SettingsTab from './app/screens/SettingsTab';
 import MyGroups from './app/screens/MyGroups';
-import PeopleNavigation from './app/screens/PeopleNavigation';
+import People from './app/screens/People';
 import CreateProfile from './app/screens/CreateProfile';
+import Profile from './app/screens/Profile';
 import { Settings, Search, Users, LayoutGrid } from 'lucide-react-native';
 
 const Stack = createNativeStackNavigator();
 const Tab = createBottomTabNavigator();
+
+// delete below screen when implemented, currently like this do we dont have empty skeleton files
+const NotificationsScreen = () => (
+  <View style={styles.tabContainer}>
+      <Text>notifications</Text>
+  </View>
+);
+
+const NotificationSettingsScreen = () => (
+  <View style={styles.tabContainer}>
+      <Text>notification settings</Text>
+  </View>
+);
+
+const ProfilePrivacyScreen = () => (
+  <View style={styles.tabContainer}>
+      <Text>profile privacy settings</Text>
+  </View>
+);
+
+const GeneralScreen = () => (
+  <View style={styles.tabContainer}>
+      <Text>general settings</Text>
+  </View>
+);
 
 const MainTabs = () => {
   return (
     <Tab.Navigator>
       <Tab.Screen name="Explore" component={Explore} options={{headerShown: false, tabBarIcon: ({ color, size }) => <Search color={color} size={size} />}} />
       <Tab.Screen name="MyGroups" component={MyGroups} options={{headerShown: false, tabBarIcon: ({ color, size }) => <LayoutGrid color={color} size={size} />}}/>
-      <Tab.Screen name="PeopleNavigation" component={PeopleNavigation} options={{headerShown: false, tabBarLabel: 'People', tabBarIcon: ({ color, size }) => <Users color={color} size={size} />}}/>
-      <Tab.Screen name="Settings" component={SettingsNavigation} options={{headerShown: false, tabBarIcon: ({ color, size }) => <Settings color={color} size={size} /> }} />
+      <Tab.Screen name="People" component={People} options={{headerShown: false, tabBarIcon: ({ color, size }) => <Users color={color} size={size} />}}/>
+      <Tab.Screen name="Settings" component={SettingsTab} options={{headerShown: false, tabBarIcon: ({ color, size }) => <Settings color={color} size={size} /> }} />
     </Tab.Navigator>
   );
 };
@@ -32,9 +58,24 @@ const App = () => {
         <Stack.Screen name="Login" component={Login} options={{ headerShown: false }} />
         <Stack.Screen name="Main" component={MainTabs} options={{ headerShown: false }} />
         <Stack.Screen name="CreateProfile" component={CreateProfile} options={{ headerShown: false }} />
+        <Stack.Screen name="Profile" component={Profile} options={{ headerTitle: ''}}/>
+        <Stack.Screen name="Notifications" component={NotificationsScreen}/>
+        <Stack.Screen name="Notification Settings" component={NotificationSettingsScreen}/>
+        <Stack.Screen name="Profile Privacy Settings" component={ProfilePrivacyScreen}/>
+        <Stack.Screen name="General Settings" component={GeneralScreen}/>
       </Stack.Navigator>
     </NavigationContainer>
   );
 };
 
 export default App;
+
+// delete this too when settings screen are implemented
+const styles = StyleSheet.create({
+  tabContainer: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#fff',
+  },
+});
